backend/requirements.txt
bcrypt==4.0.1
passlib==1.7.4
fastapi
uvicorn
beanie
motor
pydantic
pydantic-settings
python-jose[cryptography]
python-multipart
yfinance
apscheduler
aiosmtplib
python-dotenv
email-validator
requests
beautifulsoup4
slowapi
google-generativeai>=0.8.3
feedparser
